+++ beetsplug/lyrics.py
@@ -36,6 +36,7 @@ from unidecode import unidecode
 
 from beets import plugins, ui
 from beets.autotag.distance import string_dist
+from beets.dbcore import types
 from beets.util import unique_list
 from beets.util.config import sanitize_choices
 
@@ -59,7 +60,6 @@ if TYPE_CHECKING:
     )
 
 INSTRUMENTAL_LYRICS = "[Instrumental]"
-SYNCED_LYRICS_PAT = re.compile(r"\[\d\d:\d\d.\d\d\]")
 
 
 class CaptchaError(requests.exceptions.HTTPError):
@@ -227,6 +227,79 @@ class LyricsRequestHandler(RequestHandler):
             self.warn("Request error: {}", exc)
 
 
+@dataclass
+class Lyrics:
+    TRANSLATION_PAT = re.compile(r" / [^\n]+")
+    LINE_PARTS_RE = re.compile(r"^(\[\d\d:\d\d.\d\d\]|) *(.*)$")
+
+    text: str
+    backend: str
+    url: str | None = None
+    language: str | None = None
+    translation_language: str | None = None
+
+    def __post_init__(self) -> None:
+        if not self.language and self.text:
+            with suppress(langdetect.LangDetectException):
+                self.language = langdetect.detect(self.original_text).upper()
+
+    @classmethod
+    def from_item(cls, item: Item) -> Lyrics:
+        data = {"text": item.lyrics}
+        for key in ("backend", "url", "language", "translation_language"):
+            data[key] = item.get(f"lyrics_{key}", with_album=False)
+
+        return cls(**data)
+
+    @property
+    def original_text(self) -> str:
+        """Return the original text without translations."""
+        # Remove translations from the lyrics text.
+        return self.TRANSLATION_PAT.sub("", self.text).strip()
+
+    @cached_property
+    def _split_lines(self) -> list[tuple[str, str]]:
+        """Append translations to the given lyrics texts.
+
+        Lines may contain timestamps from LRCLib which need to be temporarily
+        removed for the translation. They can take any of these forms:
+                        - empty
+        Text            - text only
+        [00:00.00]      - timestamp only
+        [00:00.00] Text - timestamp with text
+        """
+        return [
+            (m[1], m[2]) if (m := self.LINE_PARTS_RE.match(line)) else ("", "")
+            for line in self.text.splitlines()
+        ]
+
+    @cached_property
+    def timestamps(self) -> list[str]:
+        return [ts for ts, _ in self._split_lines]
+
+    @cached_property
+    def text_lines(self) -> list[str]:
+        return [ln for _, ln in self._split_lines]
+
+    @property
+    def synced(self) -> bool:
+        return any(self.timestamps)
+
+    @property
+    def translated(self) -> bool:
+        return bool(self.translation_language)
+
+    def append_translations(self, translations: list[str]) -> None:
+        text_pairs = list(zip(self.text_lines, translations))
+
+        # only add the separator for non-empty translations
+        texts = [" / ".join(unique_list(filter(None, p))) for p in text_pairs]
+        # only add the space between non-empty timestamps and texts
+        self.text = "\n".join(
+            " ".join(filter(None, p)) for p in zip(self.timestamps, texts)
+        )
+
+
 class BackendClass(type):
     @property
     def name(cls) -> str:
@@ -243,7 +316,7 @@ class Backend(LyricsRequestHandler, metaclass=BackendClass):
 
     def fetch(
         self, artist: str, title: str, album: str, length: int
-    ) -> tuple[str, str] | None:
+    ) -> Lyrics | None:
         raise NotImplementedError
 
 
@@ -269,7 +342,7 @@ class LRCLyrics:
         cls, duration: float, lyrics: str | None
     ) -> str | None:
         if lyrics and (
-            m := Translator.LINE_PARTS_RE.match(lyrics.splitlines()[-1])
+            m := Lyrics.LINE_PARTS_RE.match(lyrics.splitlines()[-1])
         ):
             ts, _ = m.groups()
             if ts:
@@ -371,7 +444,7 @@ class LRCLib(Backend):
 
     def fetch(
         self, artist: str, title: str, album: str, length: int
-    ) -> tuple[str, str] | None:
+    ) -> Lyrics | None:
         """Fetch lyrics text for the given song data."""
         evaluate_item = partial(LRCLyrics.make, target_duration=length)
 
@@ -379,7 +452,9 @@ class LRCLib(Backend):
             candidates = [evaluate_item(item) for item in group]
             if item := self.pick_best_match(candidates):
                 lyrics = item.get_text(self.config["synced"].get(bool))
-                return lyrics, f"{self.GET_URL}/{item.id}"
+                return Lyrics(
+                    lyrics, self.__class__.name, f"{self.GET_URL}/{item.id}"
+                )
 
         return None
 
@@ -407,7 +482,7 @@ class MusiXmatch(Backend):
     def build_url(cls, *args: str) -> str:
         return cls.URL_TEMPLATE.format(*map(cls.encode, args))
 
-    def fetch(self, artist: str, title: str, *_) -> tuple[str, str] | None:
+    def fetch(self, artist: str, title: str, *_) -> Lyrics | None:
         url = self.build_url(artist, title)
 
         html = self.get_text(url)
@@ -429,7 +504,7 @@ class MusiXmatch(Backend):
         # sometimes there are non-existent lyrics with some content
         if "Lyrics | Musixmatch" in lyrics:
             return None
-        return lyrics, url
+        return Lyrics(lyrics, self.__class__.name, url)
 
 
 class Html:
@@ -532,13 +607,13 @@ class SearchBackend(SoupMixin, Backend):
             if check_match(candidate):
                 yield candidate
 
-    def fetch(self, artist: str, title: str, *_) -> tuple[str, str] | None:
+    def fetch(self, artist: str, title: str, *_) -> Lyrics | None:
         """Fetch lyrics for the given artist and title."""
         for result in self.get_results(artist, title):
             if (html := self.get_text(result.url)) and (
                 lyrics := self.scrape(html)
             ):
-                return lyrics, result.url
+                return Lyrics(lyrics, self.__class__.name, result.url)
 
         return None
 
@@ -751,11 +826,7 @@ class Google(SearchBackend):
 @dataclass
 class Translator(LyricsRequestHandler):
     TRANSLATE_URL = "https://api.cognitive.microsofttranslator.com/translate"
-    LINE_PARTS_RE = re.compile(r"^(\[\d\d:\d\d.\d\d\]|) *(.*)$")
     SEPARATOR = " | "
-    remove_translations = staticmethod(
-        partial(re.compile(r" / [^\n]+").sub, "")
-    )
 
     _log: Logger
     api_key: str
@@ -777,7 +848,7 @@ class Translator(LyricsRequestHandler):
             [x.upper() for x in from_languages or []],
         )
 
-    def get_translations(self, texts: Iterable[str]) -> list[tuple[str, str]]:
+    def get_translations(self, texts: Iterable[str]) -> list[str]:
         """Return translations for the given texts.
 
         To reduce the translation 'cost', we translate unique texts, and then
@@ -795,37 +866,9 @@ class Translator(LyricsRequestHandler):
         translated_text = data[0]["translations"][0]["text"]
         translations = translated_text.split(self.SEPARATOR)
         trans_by_text = dict(zip(unique_texts, translations))
-        return list(zip(texts, (trans_by_text.get(t, "") for t in texts)))
+        return [trans_by_text.get(t, "") for t in texts]
 
-    @classmethod
-    def split_line(cls, line: str) -> tuple[str, str]:
-        """Split line to (timestamp, text)."""
-        if m := cls.LINE_PARTS_RE.match(line):
-            return m[1], m[2]
-
-        return "", ""
-
-    def append_translations(self, lines: Iterable[str]) -> list[str]:
-        """Append translations to the given lyrics texts.
-
-        Lines may contain timestamps from LRCLib which need to be temporarily
-        removed for the translation. They can take any of these forms:
-                        - empty
-        Text            - text only
-        [00:00:00]      - timestamp only
-        [00:00:00] Text - timestamp with text
-        """
-        # split into [(timestamp, text), ...]]
-        ts_and_text = list(map(self.split_line, lines))
-        timestamps = [ts for ts, _ in ts_and_text]
-        text_pairs = self.get_translations([ln for _, ln in ts_and_text])
-
-        # only add the separator for non-empty translations
-        texts = [" / ".join(unique_list(filter(None, p))) for p in text_pairs]
-        # only add the space between non-empty timestamps and texts
-        return [" ".join(filter(None, p)) for p in zip(timestamps, texts)]
-
-    def translate(self, new_lyrics: str, old_lyrics: str) -> str:
+    def translate(self, lyrics: Lyrics, old_lyrics: Lyrics) -> Lyrics:
         """Translate the given lyrics to the target language.
 
         Check old lyrics for existing translations and return them if their
@@ -834,38 +877,34 @@ class Translator(LyricsRequestHandler):
 
         If the lyrics are already in the target language or not in any of
         of the source languages (if configured), they are returned as is.
-
-        The footer with the source URL is preserved, if present.
         """
         if (
-            " / " in old_lyrics
-            and self.remove_translations(old_lyrics) == new_lyrics
-        ):
+            lyrics.original_text
+        ) == old_lyrics.original_text and old_lyrics.translated:
             self.info("🔵 Translations already exist")
             return old_lyrics
 
-        lyrics_language = langdetect.detect(new_lyrics).upper()
-        if lyrics_language == self.to_language:
+        if (lyrics_language := lyrics.language) == self.to_language:
             self.info(
                 "🔵 Lyrics are already in the target language {.to_language}",
                 self,
             )
-            return new_lyrics
-
-        if self.from_languages and lyrics_language not in self.from_languages:
+        elif (
+            from_lang_config := self.from_languages
+        ) and lyrics_language not in from_lang_config:
             self.info(
-                "🔵 Configuration {.from_languages} does not permit translating"
-                " from {}",
-                self,
+                "🔵 Configuration {} does not permit translating from {}",
+                from_lang_config,
                 lyrics_language,
             )
-            return new_lyrics
+        else:
+            with self.handle_request():
+                translations = self.get_translations(lyrics.text_lines)
+                lyrics.append_translations(translations)
+                self.info("🟢 Translated lyrics to {.to_language}", self)
+                lyrics.translation_language = self.to_language
 
-        lyrics, *url = new_lyrics.split("\n\nSource: ")
-        with self.handle_request():
-            translated_lines = self.append_translations(lyrics.splitlines())
-            self.info("🟢 Translated lyrics to {.to_language}", self)
-            return "\n\nSource: ".join(["\n".join(translated_lines), *url])
+        return lyrics
 
 
 @dataclass
@@ -962,6 +1001,12 @@ class RestFiles:
 class LyricsPlugin(LyricsRequestHandler, plugins.BeetsPlugin):
     BACKEND_BY_NAME: ClassVar[dict[str, type[Backend]]] = {
         b.name: b for b in [LRCLib, Google, Genius, Tekstowo, MusiXmatch]
+    }
+    item_types: ClassVar[dict[str, types.Type]] = {
+        "lyrics_url": types.STRING,
+        "lyrics_backend": types.STRING,
+        "lyrics_language": types.STRING,
+        "lyrics_translation_language": types.STRING,
     }
 
     @cached_property
@@ -1077,18 +1122,15 @@ class LyricsPlugin(LyricsRequestHandler, plugins.BeetsPlugin):
         for item in task.imported_items():
             self.add_item_lyrics(item, False)
 
-    def find_lyrics(self, item: Item) -> str:
+    def find_lyrics(self, item: Item) -> Lyrics | None:
         album, length = item.album, round(item.length)
         matches = (
-            [
-                lyrics
-                for t in titles
-                if (lyrics := self.get_lyrics(a, t, album, length))
-            ]
+            self.get_lyrics(a, t, album, length)
             for a, titles in search_pairs(item)
+            for t in titles
         )
 
-        return "\n\n---\n\n".join(next(filter(None, matches), []))
+        return next(filter(None, matches), None)
 
     def add_item_lyrics(self, item: Item, write: bool) -> None:
         """Fetch and store lyrics for a single item. If ``write``, then the
@@ -1101,35 +1143,41 @@ class LyricsPlugin(LyricsRequestHandler, plugins.BeetsPlugin):
             self.info("🔵 Lyrics already present: {}", item)
             return
 
-        if lyrics := self.find_lyrics(item):
+        existing_lyrics = Lyrics.from_item(item)
+        if new_lyrics := self.find_lyrics(item):
             self.info("🟢 Found lyrics: {}", item)
             if translator := self.translator:
-                lyrics = translator.translate(lyrics, item.lyrics)
+                new_lyrics = translator.translate(new_lyrics, existing_lyrics)
+
+            synced_mode = self.config["synced"].get(bool)
+            if synced_mode and existing_lyrics.synced and not new_lyrics.synced:
+                self.info(
+                    "🔴 Not updating synced lyrics with non-synced ones: {}",
+                    item,
+                )
+                return
+
+            for key in ("backend", "url", "language", "translation_language"):
+                if value := getattr(new_lyrics, key, None):
+                    item[f"lyrics_{key}"] = value
+
+            lyrics_text = new_lyrics.text
         else:
             self.info("🔴 Lyrics not found: {}", item)
-            lyrics = self.config["fallback"].get()
+            lyrics_text = self.config["fallback"].get()
 
-        has_new_lyrics = lyrics not in {None, item.lyrics}
-        synced_mode = self.config["synced"].get(bool)
-        existing_synced = bool(SYNCED_LYRICS_PAT.search(item.lyrics))
-        new_synced = bool(SYNCED_LYRICS_PAT.search(lyrics or ""))
-        if has_new_lyrics and not (
-            synced_mode and existing_synced and not new_synced
-        ):
-            item.lyrics = lyrics
+        if lyrics_text not in {None, item.lyrics}:
+            item.lyrics = lyrics_text
             if write:
                 item.try_write()
             item.store()
 
-    def get_lyrics(self, artist: str, title: str, *args) -> str | None:
-        """Fetch lyrics, trying each source in turn. Return a string or
-        None if no lyrics were found.
-        """
+    def get_lyrics(self, artist: str, title: str, *args) -> Lyrics | None:
+        """Get first found lyrics, trying each source in turn."""
         self.info("Fetching lyrics for {} - {}", artist, title)
         for backend in self.backends:
             with backend.handle_request():
                 if lyrics_info := backend.fetch(artist, title, *args):
-                    lyrics, url = lyrics_info
-                    return f"{lyrics}\n\nSource: {url}"
+                    return lyrics_info
 
         return None